dev-util/codeblocks: rm '-fno-delete-null-pointer-checks' flag
Remove '-fno-delete-null-pointer-checks' compiler flag.
This flag was added to allow to run Codeblocks after compiling codeblocks-16.01
with gcc-6.1 [1] and this issue was fixed in upstream on  2016-07-16 in ticket 10875 [2].
On the 2017-03-24 in upstream ticket 11033 [3] this flag was removed as no needed anymore:
'linux: Partially Revert rev 10875 because we no longer do null pointer checks'.

Refs:
[1] http://forums.codeblocks.org/index.php?topic=21283.0
[2] https://sourceforge.net/p/codeblocks/code/10875/
[3] https://sourceforge.net/p/codeblocks/code/11033/