sci-libs/cantera: reduce 'if python' conditional choose
Reduce 'if python' conditional choose inside the 'if cti' conditional block
by replacing the 'if-else' block with introducing of
'local scons_python=$(usex python full minimal)' variable.

Thanks to @grknight for this improvement.

This changes are not implimented within 'cantera-2.3.0.ebuild'
due to the differences of the 'python3_package' variable valid values
within the mentioned block.