sci-libs/cantera: 2.4.0-r1. Enable build with sci-libs/sundials-3.2.x
This patch moves early applied 'sed' patches from src_prepare() to external
patch-file and allows to build Cantera 2.4.0 with sci-libs/sundials-3.2.x
since this version of Sundials is compatible with Sundials 3.1
and doesn't require to rebuild Cantera after it's upgrade from Sundials 3.1
to Sundials 3.2.

It was tested that Cantera successfuly build with Sundials-3.2.1
(that not in portage tree yet) and internal tests passed after "ebuild compile"
with Sundials 3.1.2 and "ebuild test" phase after intallation
of the Sundials 3.2.1 into system.

Note that Cantera 2.4.0 isn't compatible with Sundials 4.y branch
so there is a restriction for RDEPEND sci-libs/sundials-4.0.0